+++ src/test/java/org/arkecosystem/client/ConnectionManagerTest.java
@@ -1,7 +1,6 @@
 package org.arkecosystem.client;
 
-import org.arkecosystem.client.api.one.One;
-import org.arkecosystem.client.api.two.Two;
+import org.arkecosystem.client.api.Api;
 import org.junit.jupiter.api.Test;
 
 import java.util.HashMap;
@@ -15,7 +14,7 @@ public class ConnectionManagerTest {
     public void connect() {
         HashMap<String, Object> map = new HashMap<>();
         map.put("host", "dummy");
-        map.put("version", 1);
+        map.put("version", 2);
 
         ConnectionManager manager = new ConnectionManager();
         manager.connect(map);
@@ -26,7 +25,7 @@ public class ConnectionManagerTest {
     public void disconnect() {
         HashMap<String, Object> map = new HashMap<>();
         map.put("host", "dummy");
-        map.put("version", 1);
+        map.put("version", 2);
 
         ConnectionManager manager = new ConnectionManager();
         manager.connect(map);
@@ -35,16 +34,17 @@ public class ConnectionManagerTest {
         assertEquals(0, manager.getConnections().size());
     }
 
+    @Test
     public void connection() {
         HashMap<String, Object> map = new HashMap<>();
         map.put("host", "dummy");
-        map.put("version", 1);
+        map.put("version", 2);
 
         ConnectionManager manager = new ConnectionManager();
         manager.connect(map);
-        Connection<One> connection = manager.connection("main");
+        Connection<Api> connection = manager.connection("main");
         assertNotNull(connection);
-        assertEquals(One.class, connection.api().getClass());
+        assertEquals(Api.class, connection.api().getClass());
     }
 
     @Test
@@ -68,16 +68,16 @@ public class ConnectionManagerTest {
 
         HashMap<String, Object> map = new HashMap<>();
         map.put("host", "dummy");
-        map.put("version", 1);
+        map.put("version", 2);
 
-        Connection<One> connection1 = manager.connect(map);
+        Connection<Api> connection1 = manager.connect(map);
         assertNotNull(connection1);
 
         map = new HashMap<>();
         map.put("host", "dummy");
         map.put("version", 2);
 
-        Connection<Two> connection2 = manager.connect(map, "backup");
+        Connection<Api> connection2 = manager.connect(map, "backup");
         assertNotNull(connection2);
 
         assertEquals(2, manager.getConnections().size());
