+++ src/test/java/org/arkecosystem/client/api/TransactionsTest.java
@@ -1,4 +1,4 @@
-package org.arkecosystem.client.api.two;
+package org.arkecosystem.client.api;
 
 import com.google.gson.internal.LinkedTreeMap;
 import org.arkecosystem.client.Connection;
@@ -14,49 +14,49 @@ public class TransactionsTest {
 
     @Test
     void all() throws IOException {
-        Connection<Two> connection = MockHelper.connection(2);
+        Connection<Api> connection = MockHelper.connection(2);
         LinkedTreeMap<String, Object> actual = connection.api().transactions.all();
         assertTrue((boolean) actual.get("success"));
     }
 
     @Test
     void create() throws IOException {
-        Connection<Two> connection = MockHelper.connection(2);
+        Connection<Api> connection = MockHelper.connection(2);
         LinkedTreeMap<String, Object> actual = connection.api().transactions.create(new HashMap<>());
         assertTrue((boolean) actual.get("success"));
     }
 
     @Test
     void show() throws IOException {
-        Connection<Two> connection = MockHelper.connection(2);
+        Connection<Api> connection = MockHelper.connection(2);
         LinkedTreeMap<String, Object> actual = connection.api().transactions.show("dummy");
         assertTrue((boolean) actual.get("success"));
     }
 
     @Test
     void allUnconfirmed() throws IOException {
-        Connection<Two> connection = MockHelper.connection(2);
+        Connection<Api> connection = MockHelper.connection(2);
         LinkedTreeMap<String, Object> actual = connection.api().transactions.allUnconfirmed();
         assertTrue((boolean) actual.get("success"));
     }
 
     @Test
     void showUnconfirmed() throws IOException {
-        Connection<Two> connection = MockHelper.connection(2);
+        Connection<Api> connection = MockHelper.connection(2);
         LinkedTreeMap<String, Object> actual = connection.api().transactions.showUnconfirmed("dummy");
         assertTrue((boolean) actual.get("success"));
     }
 
     @Test
     void search() throws IOException {
-        Connection<Two> connection = MockHelper.connection(2);
+        Connection<Api> connection = MockHelper.connection(2);
         LinkedTreeMap<String, Object> actual = connection.api().transactions.search(new HashMap<>());
         assertTrue((boolean) actual.get("success"));
     }
 
     @Test
     void types() throws IOException {
-        Connection<Two> connection = MockHelper.connection(2);
+        Connection<Api> connection = MockHelper.connection(2);
         LinkedTreeMap<String, Object> actual = connection.api().transactions.types();
         assertTrue((boolean) actual.get("success"));
     }
