--- src/test/java/org/arkecosystem/client/api/one/SignaturesTest.java
@@ -1,21 +1,0 @@
-package org.arkecosystem.client.api.one;
-
-import com.google.gson.internal.LinkedTreeMap;
-import org.arkecosystem.client.Connection;
-import org.arkecosystem.client.MockHelper;
-import org.junit.jupiter.api.Test;
-
-import java.io.IOException;
-
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-public class SignaturesTest {
-
-    @Test
-    public void fee() throws IOException {
-        Connection<One> connection = MockHelper.connection();
-        LinkedTreeMap<String, Object> actual = connection.api().signatures.fee();
-        assertTrue((boolean) actual.get("success"));
-    }
-
-}
